New transaction id return value.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/nomis/NomisTransactionTransferResponse.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.nomis
 data class NomisTransactionTransferResponse(
   val debitTransaction: DebitTransaction,
   val creditTransaction: CreditTransaction,
-  val transactionId: Int,
+  val transactionId: Long,
 )
 
 data class DebitTransaction(
